requirements.txt
Django~=5.1.2
djangorestframework~=3.15.2
celery~=5.4.0

requests~=2.32.3
